ARXIVNG-109 refactor to separate advanced search
